refac(salesforce) refactor zapier client
